+++ packages/platform/src/contracts/error/index.ts
@@ -1,0 +1,7 @@
+export * from "./ErrorCategory";
+export * from "./ErrorCode";
+export * from "./ErrorContext";
+export * from "./ErrorDetails";
+export * from "./ErrorSeverity";
+export * from "./PlatformError";
+export * from "./PlatformResult";